requirements.txt
aiogram==3.20.0
tortoise-orm==0.20.0
python-dotenv==1.0.1
aiosqlite==0.19.0